refactor: 기타 누락된 부분 수정
