+++ src/main/java/Application.java
@@ -21,7 +21,7 @@ public class Application {
 			new PaymentService(tableRepository), new TableOrderService(tableRepository, menuRepository));
 
 		PageController pageController = new PageController(inputView, chickenController);
-		while (pageController.runIfNotExit())
-			;
+		while (pageController.runIfNotExit()) {
+		}
 	}
 }
